HW-0000 Travel agency service
- reworked beforeSaveBooking method as decorator
